refactor(catalog): move catalog serializers to V2 namespace
## Context

The catalog serializers served /api/v2 exclusively but still lived
under V1::. The convention is that everything on /api/v2 serializes
through V2::, alongside the rate phase, rate override, contract and
catalog plan serializers already there.

## Description

Move the Product, ProductCategory, ProductFilter, RateCard,
RateCardRate and PlanRateCard serializers from V1:: to V2::, repointing
the v2 controllers. Consolidate CatalogPlan onto V2::CatalogPlanSerializer:
the plan.created / plan.updated webhooks now serialize through it, so
their payload gains applied_rate_cards_count and drops updated_at.
Utils::ActivityLog#object_serialized resolves V1:: first, so legacy
records keep today's shape, then falls through to V2:: for catalog-only
classes, keeping catalog activity logs producing payloads.
