+++ app/controllers/api/v2/rate_cards_controller.rb
@@ -75,7 +75,7 @@ module Api
           render(
             json: ::CollectionSerializer.new(
               result.rate_cards.includes(:product, :product_filter, :rates),
-              ::V1::RateCardSerializer,
+              ::V2::RateCardSerializer,
               collection_name: "rate_cards",
               meta: pagination_metadata(result.rate_cards)
             )
@@ -146,7 +146,7 @@ module Api
       end
 
       def render_rate_card(rate_card)
-        render(json: ::V1::RateCardSerializer.new(rate_card, root_name: "rate_card", includes: %i[active_rate taxes]))
+        render(json: ::V2::RateCardSerializer.new(rate_card, root_name: "rate_card", includes: %i[active_rate taxes]))
       end
 
       def resource_name
