+++ app/serializers/v2/plan_rate_card_serializer.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module V1
+module V2
   class PlanRateCardSerializer < ModelSerializer
     def serialize
       {
