+++ app/serializers/v2/product_category_serializer.rb
@@ -1,15 +1,15 @@
 # frozen_string_literal: true
 
-module V1
-  class CatalogPlanSerializer < ModelSerializer
+module V2
+  class ProductCategorySerializer < ModelSerializer
     def serialize
       {
         lago_id: model.id,
         name: model.name,
-        invoice_display_name: model.invoice_display_name,
         code: model.code,
         description: model.description,
-        currency: model.currency,
+        invoice_display_name: model.invoice_display_name,
+        products_count: model.products.size,
         created_at: model.created_at.iso8601,
         updated_at: model.updated_at.iso8601
       }
