+++ app/serializers/v2/rate_card_rate_serializer.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module V1
+module V2
   class RateCardRateSerializer < ModelSerializer
     def serialize
       {
