+++ app/services/webhooks/plans/updated_service.rb
@@ -8,7 +8,7 @@ module Webhooks
       def object_serializer
         # A catalog plan is a plan to the end user: same webhook type, its own
         # (leaner) payload.
-        return ::V1::CatalogPlanSerializer.new(object, root_name: "plan") if object.is_a?(CatalogPlan)
+        return ::V2::CatalogPlanSerializer.new(object, root_name: "plan") if object.is_a?(CatalogPlan)
 
         ::V1::PlanSerializer.new(
           object,
